+++ src/data_preprocessing/data_contraints.jl
@@ -12,7 +12,9 @@ end
 
 If nothing is supplied, all features are assumed to be mutable in both directions.
 """
-function mutability_constraints(counterfactual_data::CounterfactualData, mutability::Nothing)
+function mutability_constraints(
+    counterfactual_data::CounterfactualData, mutability::Nothing
+)
     return [:both for i in 1:size(counterfactual_data.X)[1]]
 end
 
@@ -21,7 +23,9 @@ mutability_constraints(counterfactual_data::CounterfactualData, mutability::Vect
 
 If `mutability` is already a vector of symbols, it is returned as is. 
 """
-function mutability_constraints(counterfactual_data::CounterfactualData, mutability::Vector{Symbol})
+function mutability_constraints(
+    counterfactual_data::CounterfactualData, mutability::Vector{Symbol}
+)
     return mutability
 end
 
@@ -30,7 +34,9 @@ mutability_constraints(counterfactual_data::CounterfactualData, mutability::Vect
 
 If `mutability` is already a vector of integers, these are assumed to be indices of immutable features. All other features are assumed to be mutable in both directions. This offers a convenient way to quickly specify immutable features.
 """
-function mutability_constraints(counterfactual_data::CounterfactualData, mutability::Vector{Int})
+function mutability_constraints(
+    counterfactual_data::CounterfactualData, mutability::Vector{Int}
+)
     mutability_final = mutability_constraints(counterfactual_data, nothing) # starting point
     for i in mutability
         # Set all referenced indices to immutable:
@@ -39,21 +45,26 @@ function mutability_constraints(counterfactual_data::CounterfactualData, mutabil
     return mutability_final
 end
 
-
 """
     mutability_constraints(counterfactual_data::CounterfactualData, mutability::Pair{K,V}...) where {K<:Int,V<:Symbol}
 
 If pairs of feature indices (`::Int`) and constraints (`::Symbol`) are supplied, the given constraints are applied to the corresponding features. All other features are assumed to be mutable in both directions.
 """
-function mutability_constraints(counterfactual_data::CounterfactualData, mutability::Pair{K,V}...) where {K<:Int,V<:Symbol}
+function mutability_constraints(
+    counterfactual_data::CounterfactualData, mutability::Pair{K,V}...
+) where {K<:Int,V<:Symbol}
     mutability_final = mutability_constraints(counterfactual_data, nothing) # starting point
-    for (k,v) in mutability
+    for (k, v) in mutability
         mutability_final[k] = v
     end
     return mutability_final
 end
 
-mutability_constraints(counterfactual_data::CounterfactualData, mutability::Tuple{Vararg{Pair{Int, Symbol}}}) = mutability_constraints(counterfactual_data, mutability...)
+function mutability_constraints(
+    counterfactual_data::CounterfactualData, mutability::Tuple{Vararg{Pair{Int,Symbol}}}
+)
+    mutability_constraints(counterfactual_data, mutability...)
+end
 
 """
     mutability_constraints!(counterfactual_data::CounterfactualData, mutability)