+++ src/generators/Generators.jl
@@ -84,11 +84,12 @@ end
 
 Computes the total loss of a counterfactual explanation with respect to the search objective.
 """
-total_loss(ce::AbstractCounterfactualExplanation) =
+function total_loss(ce::AbstractCounterfactualExplanation)
     if hasfield(typeof(ce.generator), :loss)
         ℓ(ce.generator, ce) + h(ce.generator, ce)
     else
         nothing
     end
+end
 
 end