+++ src/objectives/traits.jl
@@ -21,8 +21,9 @@ needs_neighbours(::NeedsNeighbours, x) = true
 
 Check if a generator needs access to neighbors in the target class.
 """
-needs_neighbours(gen::AbstractGenerator) =
+function needs_neighbours(gen::AbstractGenerator)
     hasfield(typeof(gen), :penalty) ? any(needs_neighbours.(gen.penalty)) : false
+end
 
 """
     needs_neighbours(ce::AbstractCounterfactualExplanation)