+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 *Note*: We try to adhere to these practices as of version [v1.1.1].
 
-## Version [1.4.6] - 2025-01-14
+## Version [1.4.6] - 2025-07-25
 
 ### Added